Struct vec2d::Vec2D

pub struct Vec2D<T> {
    // some fields omitted
}

Container for 2D data

Methods

impl<Elem: Copy> Vec2D<Elem>

fn from_example(size: Size, example: &Elem) -> Vec2D<Elem>

Create a Vec2D with the given size. All elements are initialized as copies of the example element.

fn from_vec(size: Size, src: Vec<Elem>) -> Option<Vec2D<Elem>>

Create a Vec2D with the given size. The contents are set to src. None is returned if the size does not match the length of src.

fn get(&self, coord: Coord) -> Option<&Elem>

Returns element at the given coord or None if the coord is outside the Vec2D

Example

let v = Vec2D::from_vec (
  Size { width: 3, height: 3 },
  vec!['a','b','c','d','e','f','g','h','i']
).unwrap();
assert_eq!(v.get (Coord { x: 1, y: 0 }), Some(&'b'));
assert_eq!(v.get (Coord { x: 1, y: 2 }), Some(&'h'));
assert_eq!(v.get (Coord { x: 3, y: 0 }), None);

fn get_mut(&mut self, coord: Coord) -> Option<&mut Elem>

Returns a mutable reference to the element at the given coord or None if the coord is outside the Vec2D

Example

let mut v = Vec2D::from_vec (
  Size { width: 3, height: 3 },
  vec!['a','b','c','d','e','f','g','h','i']
).unwrap();
assert_eq!(v.get_mut (Coord { x: 1, y: 0 }), Some(&mut 'b'));
assert_eq!(v.get_mut (Coord { x: 1, y: 2 }), Some(&mut 'h'));
assert_eq!(v.get_mut (Coord { x: 3, y: 0 }), None);

fn rect(&self) -> Rect

Shortcut for self.size.rect()

fn resize(&mut self, new_size: Size, value: Elem)

Resize in-place so that size() is equal to new_size

fn size(&self) -> Size

Width and height

fn iter<'a>(&'a self) -> RectIter<'a, Elem>

Iterator over the entire Vec2D.

fn rect_iter<'a>(&'a self, rect: Rect) -> Option<RectIter<'a, Elem>>

Create an iterator over a rectangular region of the Vec2D. None is returned if the given rect does not fit entirely within the Vec2D.

fn rect_iter_at<'a>(&'a self, rect: Rect, start: Coord) -> Option<RectIter<'a, Elem>>

Create an iterator over a rectangular region of the Vec2D with the start coord. None is returned if the given rect does not fit entirely within the Vec2D or if the start coord is not within rect.

fn iter_mut<'a>(&'a mut self) -> RectIterMut<'a, Elem>

Mutable iterater over the entire Vec2D.

fn rect_iter_mut<'a>(&'a mut self, rect: Rect) -> Option<RectIterMut<'a, Elem>>

Create a mutable iterator over a rectangular region of the Vec2D. None is returned if the given rect does not fit entirely within the Vec2D.

fn rect_iter_mut_at<'a>(&'a mut self, rect: Rect, start: Coord) -> Option<RectIterMut<'a, Elem>>

Create a mutable iterator over a rectangular region of the Vec2D with the start coord. None is returned if the given rect does not fit entirely within the Vec2D or if the start coord is not within rect.